feat: final refactor on documenation & project structure
