+++ src/config/routes.ts
@@ -1,0 +1,6 @@
+export const base = "/";
+
+export default {
+  rides: `${base}rides`,
+  healthCheck: `${base}health`,
+};
